+++ src/main/java/fr/siamois/repositories/DocumentRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DocumentRepository extends CrudRepository<Document, Integer> {
+public interface DocumentRepository extends CrudRepository<Document, Long> {
 
     @Query(
             nativeQuery = true,
